requirements.txt
scikit-learn==1.3.1
scikit-learn-intelex==2023.2.1
matplotlib==3.6.2
pandas==2.1.3
joblibspark==0.5.2
xgboost==2.0.2
findspark==2.0.1
tqdm
iteration_utilities
